refactor(client): extract DiscographyTab from ArtistPage into memoized component (C26)
Move the discography TabsContent verbatim into components/artist/DiscographyTab
(React.memo). Co-locate the pure getSongsByRelease helper in the tab and
consume shared badge/color helpers from artistPageUtils. Also drop the
dead getTotalMarketingCost helper (unused). No behavior change.
